+++ db/migrations/9_create_nextval_with_xact_lock_extension.sql
@@ -1,1 +1,0 @@
-CREATE EXTENSION nextval_with_xact_lock;